fix: set inventoryReserved correctly on SimpleInventory insert
